requirements.txt
requests>=2.28.0
PyYAML>=6.0
loguru>=0.7.2
rich>=13.0.0
Flask>=2.0.0
Pillow>=10.0.0
mercantile>=1.2.1
psutil>=5.9.0
